+++ src/inet/routing/rip/RIPRouting.cc
@@ -281,7 +281,7 @@ RIPRoute *RIPRouting::importRoute(IRoute *route, RIPRoute::RouteType type, int m
     }
 
     ripRoutes.push_back(ripRoute);
-    emit(numRoutesSignal, ripRoutes.size());
+    emit(numRoutesSignal, (long)ripRoutes.size());
     return ripRoute;
 }
 
@@ -886,7 +886,7 @@ void RIPRouting::addRoute(const L3Address& dest, int prefixLength, const Interfa
     ripRoute->setLastUpdateTime(simTime());
     ripRoute->setChanged(true);
     ripRoutes.push_back(ripRoute);
-    emit(numRoutesSignal, ripRoutes.size());
+    emit(numRoutesSignal, (long)ripRoutes.size());
     triggerUpdate();
 }
 
@@ -1028,7 +1028,7 @@ void RIPRouting::purgeRoute(RIPRoute *ripRoute)
         ripRoutes.erase(end, ripRoutes.end());
     delete ripRoute;
 
-    emit(numRoutesSignal, ripRoutes.size());
+    emit(numRoutesSignal, (long)ripRoutes.size());
 }
 
 /**
@@ -1128,7 +1128,7 @@ void RIPRouting::deleteInterface(const InterfaceEntry *ie)
             it++;
     }
     if (emitNumRoutesSignal)
-        emit(numRoutesSignal, ripRoutes.size());
+        emit(numRoutesSignal, (long)ripRoutes.size());
 }
 
 int RIPRouting::getInterfaceMetric(InterfaceEntry *ie)
